ToDotNot unit test conversion - part 6.
